ubi-bao: split parser
